renamed and refactored hints package to use reponsemodel
