--- src/main/java/HintGenerator/HintGeneratorOutBoundary.java
@@ -1,8 +1,0 @@
-package HintGenerator;
-import entities.MazeCell;
-
-import java.util.List;
-
-public interface HintGeneratorOutBoundary {
-    public void displayHint(List<MazeCell> hint);
-}
