--- src/main/java/HintGenerator/HintGeneratorPresenter.java
@@ -1,11 +1,0 @@
-package HintGenerator;
-
-import entities.MazeCell;
-
-import java.util.List;
-
-public class HintGeneratorPresenter implements HintGeneratorOutBoundary {
-    public void displayHint(List<MazeCell> hint) {
-        System.out.println(hint);
-    }
-}
