+++ src/main/java/hints/HintGenerator.java
@@ -1,4 +1,4 @@
-package HintGenerator;
+package hints;
 
 import entities.*;
