+++ src/main/java/hints/HintGeneratorInBoundary.java
@@ -1,7 +1,7 @@
-package HintGenerator;
+package hints;
 
 import entities.GameState;
 
 public interface HintGeneratorInBoundary {
-    public void generateHint(GameState maze);
+    public HintResponseModel generateHint(GameState maze);
 }
