+++ src/main/java/hints/HintGeneratorOutBoundary.java
@@ -1,0 +1,5 @@
+package hints;
+
+public interface HintGeneratorOutBoundary {
+    public HintResponseModel displayHint(HintResponseModel responseModel);
+}
