+++ src/main/java/hints/HintGeneratorPresenter.java
@@ -1,0 +1,7 @@
+package hints;
+
+public class HintGeneratorPresenter implements HintGeneratorOutBoundary {
+    public HintResponseModel displayHint(HintResponseModel reponseModel) {
+        return reponseModel;
+    }
+}
